requirements.txt
groq==0.11.0
httpx==0.27.2
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0
langchain-text-splitters==0.3.0
chromadb==0.5.3
pdfplumber==0.11.1
python-dotenv==1.0.1
onnxruntime>=1.20.0
tokenizers>=0.19.1
requests>=2.32.0
tqdm>=4.66.0
sentence-transformers>=3.0.0
gradio==4.44.0